Update README
Although bitbake rebuilds acpi-tables when something is modified
related to it, these changes are not being copied to the initramfs
which in turn, is not being updated on the final image.

The clean step was not very obvious to me at first; so thought it might
be the same for others as well.

Signed-off-by: Arun Bharadwaj <arun@gumstix.com>