arduino.asl: create arduino-all.asl
Copy arduino.asl to arduino-all.asl which is a solution asl with
the following enabled:
MUX_I2C
MUX_SPI
MUX_UART_4WIRE

Remove defines from arduino.asl. This table requires to put the
defines on the iasl command line to make them under control of
the recipe or Makefile.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>