galileo: Move spi.asli include to gpiomux.asli
Currently it's not possible to enable any of SPI connected peripheral via
ConfigFS. This happens due to broken initial tables in Intel Galileo board,
because ACPI can't add _CRS on-the-fly, the gpiomux table has to be applied to
initramfs image, but this won't provide a crucial information for the SPI,
i.e. chip select GPIO resource.

Fix this by moving spi.asli include to gpiomux.asli.

Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>